+++ lib/flooding_app_design_system.dart
@@ -1,5 +1,4 @@
 library;
 
-export 'src/utils/theme.dart';
-export 'src/utils/component.dart';
-export 'src/screen_component/sign_up.dart';
+export 'src/theme.dart';
+export 'src/component.dart';