+++ CHANGELOG.md
@@ -14,3 +14,6 @@
 * Update CI
 * export files
 * add Button Components
+* add Textfield
+* add DropDown Button
+* add Icon